requirements.txt
streamlit 
scikit-learn 
beautifulsoup4 
pypdf 
google-generativeai